Optimized

+++ publish/trigger.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
 return [
     'default' => [
         'host' => env('TRIGGER_HOST', ''),
-        'port' => env('TRIGGER_PORT', 3306),
+        'port' => (int) env('TRIGGER_PORT', 3306),
         'user' => env('TRIGGER_USER', ''),
         'password' => env('TRIGGER_PASSWORD', ''),
         'databases_only' => env('TRIGGER_DATABASES_ONLY', '') ? explode(',', env('TRIGGER_DATABASES_ONLY')) : [],
